amovetile parsing and listing tfuncs o shared_utils
